musicBot2/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2025.7.21
youtube-search-python>=1.6.6
requests==2.31.0
aiofiles>=23.2.1 
PyNaCl>=1.5.0
aiohttp==3.9.1
httpx==0.24.1
httpcore==0.17.3